+++ src/seedcase_sprout/core/check_datapackage/__init__.py
@@ -5,6 +5,7 @@ from .check_error_matcher import CheckErrorMatcher
 from .check_package_properties import check_package_properties
 from .check_properties import check_properties
 from .check_resource_properties import check_resource_properties
+from .exclude_matching_errors import exclude_matching_errors
 from .required_fields import (
     PACKAGE_RECOMMENDED_FIELDS,
     PACKAGE_REQUIRED_FIELDS,
@@ -22,4 +23,5 @@ __all__ = [
     "PACKAGE_REQUIRED_FIELDS",
     "RESOURCE_REQUIRED_FIELDS",
     "RequiredFieldType",
+    "exclude_matching_errors",
 ]
