+++ src/seedcase_sprout/core/sprout_checks/check_id_in_resource_path.py
@@ -1,6 +1,6 @@
 from pathlib import Path
 
-from seedcase_sprout.core.check_datapackage.check_error import CheckError
+from seedcase_sprout.core.check_datapackage import CheckError
 from seedcase_sprout.core.sprout_checks.get_json_path_to_resource_field import (
     get_json_path_to_resource_field,
 )
