+++ src/seedcase_sprout/core/sprout_checks/check_no_inline_data.py
@@ -1,4 +1,4 @@
-from seedcase_sprout.core.check_datapackage.check_error import CheckError
+from seedcase_sprout.core.check_datapackage import CheckError
 from seedcase_sprout.core.sprout_checks.get_json_path_to_resource_field import (
     get_json_path_to_resource_field,
 )
