+++ src/seedcase_sprout/core/sprout_checks/required_fields.py
@@ -1,24 +1,21 @@
-from seedcase_sprout.core import check_datapackage
-from seedcase_sprout.core.check_datapackage.required_fields import (
-    RequiredFieldType,
-)
+import seedcase_sprout.core.check_datapackage as cdp
 
 # Sprout-specific required fields and their types
 
 PACKAGE_SPROUT_REQUIRED_FIELDS = (
-    check_datapackage.PACKAGE_REQUIRED_FIELDS
-    | check_datapackage.PACKAGE_RECOMMENDED_FIELDS
+    cdp.PACKAGE_REQUIRED_FIELDS
+    | cdp.PACKAGE_RECOMMENDED_FIELDS
     | {
-        "title": RequiredFieldType.str,
-        "description": RequiredFieldType.str,
-        "version": RequiredFieldType.str,
-        "created": RequiredFieldType.str,
+        "title": cdp.RequiredFieldType.str,
+        "description": cdp.RequiredFieldType.str,
+        "version": cdp.RequiredFieldType.str,
+        "created": cdp.RequiredFieldType.str,
     }
 )
 PACKAGE_SPROUT_REQUIRED_FIELDS.pop("resources", None)
 
-RESOURCE_SPROUT_REQUIRED_FIELDS = check_datapackage.RESOURCE_REQUIRED_FIELDS | {
-    "title": RequiredFieldType.str,
-    "description": RequiredFieldType.str,
+RESOURCE_SPROUT_REQUIRED_FIELDS = cdp.RESOURCE_REQUIRED_FIELDS | {
+    "title": cdp.RequiredFieldType.str,
+    "description": cdp.RequiredFieldType.str,
 }
 RESOURCE_SPROUT_REQUIRED_FIELDS.pop("data", None)
